+++ src/main/java/org/cryptomator/cryptofs/fh/OpenCryptoFile.java
@@ -1,12 +1,14 @@
 package org.cryptomator.cryptofs.fh;
 
+import jakarta.inject.Inject;
 import org.cryptomator.cryptofs.EffectiveOpenOptions;
 import org.cryptomator.cryptofs.ch.CleartextFileChannel;
+import org.cryptomator.cryptofs.inuse.InUseManager;
+import org.cryptomator.cryptofs.inuse.UseToken;
 import org.cryptomator.cryptolib.api.Cryptor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import jakarta.inject.Inject;
 import java.io.Closeable;
 import java.io.IOException;
 import java.nio.channels.FileChannel;
@@ -26,6 +28,7 @@ public class OpenCryptoFile implements Closeable {
 
 	private final FileCloseListener listener;
 	private final AtomicReference<Instant> lastModified;
+	private final InUseManager inUseManager;
 	private final Cryptor cryptor;
 	private final FileHeaderHolder headerHolder;
 	private final ChunkIO chunkIO;
@@ -34,11 +37,22 @@ public class OpenCryptoFile implements Closeable {
 	private final OpenCryptoFileComponent component;
 
 	private final AtomicInteger openChannelsCount = new AtomicInteger(0);
+	private volatile UseToken useToken;
 
 	@Inject
 	public OpenCryptoFile(FileCloseListener listener, Cryptor cryptor, FileHeaderHolder headerHolder, ChunkIO chunkIO, //
 						  @CurrentOpenFilePath AtomicReference<Path> currentFilePath, @OpenFileSize AtomicLong fileSize, //
-						  @OpenFileModifiedDate AtomicReference<Instant> lastModified, OpenCryptoFileComponent component) {
+						  @OpenFileModifiedDate AtomicReference<Instant> lastModified, OpenCryptoFileComponent component, //
+						  InUseManager inUseManager) {
+		this(listener, cryptor, headerHolder, chunkIO, currentFilePath, fileSize, lastModified, component, inUseManager, UseToken.CLOSED_TOKEN);
+	}
+
+
+	//for testing
+	OpenCryptoFile(FileCloseListener listener, Cryptor cryptor, FileHeaderHolder headerHolder, ChunkIO chunkIO, //
+				   @CurrentOpenFilePath AtomicReference<Path> currentFilePath, @OpenFileSize AtomicLong fileSize, //
+				   @OpenFileModifiedDate AtomicReference<Instant> lastModified, OpenCryptoFileComponent component, //
+				   InUseManager inUseManager, UseToken token) {
 		this.listener = listener;
 		this.cryptor = cryptor;
 		this.headerHolder = headerHolder;
@@ -47,6 +61,8 @@ public class OpenCryptoFile implements Closeable {
 		this.fileSize = fileSize;
 		this.component = component;
 		this.lastModified = lastModified;
+		this.inUseManager = inUseManager;
+		this.useToken = token;
 	}
 
 	/**
@@ -66,6 +82,9 @@ public class OpenCryptoFile implements Closeable {
 
 		openChannelsCount.incrementAndGet(); // synchronized context, hence we can proactively increase the number
 		try {
+			if (options.writable() && useToken.isClosed()) { //the token was closed prematurely, so we try to get a new one
+				useToken = inUseManager.use(path);
+			}
 			ciphertextFileChannel = path.getFileSystem().provider().newFileChannel(path, options.createOpenOptionsForEncryptedFile(), attrs);
 			initFileHeader(options, ciphertextFileChannel);
 			initFileSize(ciphertextFileChannel);
@@ -81,7 +100,6 @@ public class OpenCryptoFile implements Closeable {
 				closeQuietly(ciphertextFileChannel);
 			}
 		}
-
 		assert cleartextFileChannel != null; // otherwise there would have been an exception
 		chunkIO.registerChannel(ciphertextFileChannel, options.writable());
 		return cleartextFileChannel;
@@ -122,20 +140,19 @@ public class OpenCryptoFile implements Closeable {
 	 * Initialization happens at most once per open file. Subsequent invocations are no-ops.
 	 */
 	private void initFileSize(FileChannel ciphertextFileChannel) throws IOException {
-		if (fileSize.get() == -1l) {
+		if (fileSize.get() == -1L) {
 			LOG.trace("First channel for this openFile. Initializing file size...");
-			long cleartextSize = 0l;
+			long cleartextSize = 0L;
 			try {
 				long ciphertextSize = ciphertextFileChannel.size();
-				if (ciphertextSize > 0l) {
+				if (ciphertextSize > 0L) {
 					long payloadSize = ciphertextSize - cryptor.fileHeaderCryptor().headerSize();
 					cleartextSize = cryptor.fileContentCryptor().cleartextSize(payloadSize);
 				}
 			} catch (IllegalArgumentException e) {
 				LOG.warn("Invalid cipher text file size. Assuming empty file.", e);
-				assert cleartextSize == 0l;
 			}
-			fileSize.compareAndSet(-1l, cleartextSize);
+			fileSize.compareAndSet(-1L, cleartextSize);
 		}
 	}
 
@@ -144,7 +161,7 @@ public class OpenCryptoFile implements Closeable {
 	 */
 	public Optional<Long> size() {
 		long val = fileSize.get();
-		if (val == -1l) {
+		if (val == -1L) {
 			return Optional.empty();
 		} else {
 			return Optional.of(val);
@@ -165,10 +182,16 @@ public class OpenCryptoFile implements Closeable {
 
 	/**
 	 * Updates the current ciphertext file path, if it is not already set to null (i.e., the openCryptoFile is deleted)
+	 *
 	 * @param newFilePath new ciphertext path
 	 */
 	public void updateCurrentFilePath(Path newFilePath) {
-		currentFilePath.updateAndGet(p -> p == null ? null : newFilePath);
+		currentFilePath.getAndUpdate(p -> p == null ? null : newFilePath);
+		if (newFilePath != null) {
+			useToken.moveTo(newFilePath);
+		} else {
+			useToken.close(); //encrypted file will be deleted, hence we can stop checking usage
+		}
 	}
 
 	private synchronized void cleartextChannelClosed(FileChannel ciphertextFileChannel) {
@@ -183,7 +206,8 @@ public class OpenCryptoFile implements Closeable {
 	@Override
 	public void close() {
 		var p = currentFilePath.get();
-		if(p != null) {
+		if (p != null) {
+			useToken.close();
 			listener.close(p, this);
 		}
 	}