+++ src/test/java/org/cryptomator/cryptofs/CryptoFileChannelWriteReadIntegrationTest.java
@@ -9,6 +9,8 @@
 package org.cryptomator.cryptofs;
 
 import com.google.common.jimfs.Jimfs;
+import org.awaitility.Awaitility;
+import org.cryptomator.cryptofs.common.Constants;
 import org.cryptomator.cryptofs.util.ByteBuffers;
 import org.cryptomator.cryptolib.api.Masterkey;
 import org.cryptomator.cryptolib.api.MasterkeyLoader;
@@ -145,17 +147,67 @@ public class CryptoFileChannelWriteReadIntegrationTest {
 
 	@Nested
 	@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+	public class InUseFeatureActivated {
+
+		private FileSystem inMemoryFs;
+		private FileSystem fileSystem;
+
+		private Path file;
+		private Path vaultPath;
+
+		@BeforeAll
+		public void beforeAll() throws IOException, MasterkeyLoadingFailedException {
+			inMemoryFs = Jimfs.newFileSystem();
+			vaultPath = inMemoryFs.getPath("vault");
+			Files.createDirectories(vaultPath);
+			MasterkeyLoader keyLoader = Mockito.mock(MasterkeyLoader.class);
+			Mockito.when(keyLoader.loadKey(Mockito.any())).thenAnswer(ignored -> new Masterkey(new byte[64]));
+			var properties = cryptoFileSystemProperties().withKeyLoader(keyLoader).withOwner("cryptobot").build();
+			CryptoFileSystemProvider.initialize(vaultPath, properties, URI.create("test:key"));
+			fileSystem = CryptoFileSystemProvider.newFileSystem(vaultPath, properties);
+			file = fileSystem.getPath("/test.txt");
+		}
+
+		@AfterAll
+		public void afterAll() throws IOException {
+			fileSystem.close();
+			inMemoryFs.close();
+		}
+
+		@Test
+		@DisplayName("Opening a file channel creates an in-use file and removes it on close")
+		public void testOpeningFCCreatesInUseFile() throws IOException {
+			try (var writer = FileChannel.open(file, CREATE, WRITE)) {
+				Awaitility.await().atLeast( 4900, TimeUnit.MILLISECONDS) //
+						.atMost(8000, TimeUnit.MILLISECONDS) //
+						.until(() -> numberOfInUseFiles() == 1);
+			}
+			var numberAfterClose = numberOfInUseFiles();
+			Assertions.assertEquals(0, numberAfterClose);
+		}
+
+		private long numberOfInUseFiles() throws IOException {
+			try (var encryptedFiles = Files.walk(vaultPath.resolve("d"))) {
+				var inUseFiles = encryptedFiles.filter(p -> p.getFileName().toString().endsWith(Constants.INUSE_FILE_SUFFIX));
+				return inUseFiles.count();
+			}
+		}
+	}
+
+	@Nested
+	@TestInstance(TestInstance.Lifecycle.PER_CLASS)
 	public class PlatformIndependent {
 
 		private FileSystem inMemoryFs;
 		private FileSystem fileSystem;
 
 		private Path file;
+		private Path vaultPath;
 
 		@BeforeAll
 		public void beforeAll() throws IOException, MasterkeyLoadingFailedException {
 			inMemoryFs = Jimfs.newFileSystem();
-			Path vaultPath = inMemoryFs.getPath("vault");
+			vaultPath = inMemoryFs.getPath("vault");
 			Files.createDirectories(vaultPath);
 			MasterkeyLoader keyLoader = Mockito.mock(MasterkeyLoader.class);
 			Mockito.when(keyLoader.loadKey(Mockito.any())).thenAnswer(ignored -> new Masterkey(new byte[64]));
@@ -175,6 +227,7 @@ public class CryptoFileChannelWriteReadIntegrationTest {
 		public void afterEach() throws IOException {
 			Files.deleteIfExists(file);
 		}
+
 
 		//https://github.com/cryptomator/cryptofs/issues/173
 		@Test