+++ CHANGELOG.md
@@ -10,7 +10,8 @@ Changes to prior versions can be found on the [Github release page](https://gith
 ## [Unreleased](https://github.com/cryptomator/cryptofs/compare/2.9.0...HEAD)
 
 ### Added
-* [changelog](CHANGELOG.md) file
+* Files-in-Use: Optional feature to indicate for external parties if an encrypted file is currently opened by this filesystem ([#312](https://github.com/cryptomator/cryptofs/pull/312))
+* Changelog file
 
 ### Changed
 * Use JDK 25 for build (bf26d6c9cd15a2489126ee0409a8ec9eca59da0c)